backend/requirements.txt
# FastAPI Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Storage
supabase==2.3.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Task Queue
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# AI Services
google-generativeai==0.8.3
agentic-doc==0.2.10
requests==2.31.0
aiohttp==3.9.1
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
numpy==1.25.2
scikit-image==0.22.0
imagehash==4.3.1
sentence-transformers==2.2.2

# WebSocket & Real-time
websockets==12.0

# Report Generation
python-pptx==0.6.23
jinja2==3.1.2
weasyprint==60.2

# File Processing
python-magic==0.4.27
aiofiles==23.2.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
loguru==0.7.2
python-dateutil==2.8.2
typing-extensions==4.8.0 